Core username/password algorithims happen here
When a new user registers a series of if statements check if it meets username & password requirements. If it meets requirements the signup() method is called. For security purposes the password is used as a character array versus a string in which case immutable and stays in memory pool, making it vulnerable to attack. The password is then salted to form a new password in a character array with the original + the salt characters. The password is then converted from a character array to a byte array which is then passed to the hashing method generateHash() and a hash string is output and then stored as the value in a Hashmap, with the username as the key.

Logging in follows a similar route of password conversions until it is output as a String hashcode and checked against the stored hashcode password.